requirements.txt
# jm_requirement
pyspark==3.5.7
mlflow==3.8.0
pytest==9.0.2
python-Levenshtein==0.27.3
pandas==2.3.3
numpy==2.4.0
boto3==1.35.0
